+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## v4.0.3-Alpha
 - Add translation for Portuguese (locale pt_PT) ([#639](https://github.com/userfrosting/UserFrosting/issues/639)) - Thanks @brunomnsilva
 - UfAlert style customization (See [#634](https://github.com/userfrosting/UserFrosting/issues/634))
-- Message stream now support raw placeholder interpolation (See [#621](https://github.com/userfrosting/UserFrosting/issues/621))
+- Message stream can now display raw placeholder using the `|raw` filter in the placeholder name. Requires latest version of the [i18n](https://github.com/userfrosting/i18n) component (See [#621](https://github.com/userfrosting/UserFrosting/issues/621))
 - Fix the "Root account" message breaking the UI on smaller screens (See [#641](https://github.com/userfrosting/UserFrosting/issues/641)) - Thanks @brunomnsilva !
 
 ## v4.0.2-Alpha